Upgrade to Stripe PaymentElement This upgrade requires a clientSecret to render the PaymentElement We now have to create an incomplete subscription in order to render the checkout page. The subscription is also completed on that page and the user is redirected on completion.
We now redirect the user to a Stripe controller to verify the
Subscription and make sure it's synced to the database.

We've also implemented some tweaks for metered billing so that users can
subscribe for $0.008/hour/server, for example. Your application will be
responsible for reporting usage to Stripe. We've added an upcoming
invoice view so that users can see their usage and billing for the next
billing period.