Correct example of `record` param for Omniauth (#755)
The Jumpstart::Omniauth::Callback module only accepts signed GlobalIDs that were created with an explicit purpose of `:oauth`, but the single example in the docs of creating a signed GlobalID does not include that parameter, so the example doesn't actually work. This change adds the necessary parameter to the example so that the example could potentially work.